+++ src/components/base/Button/Button.tsx
@@ -1,7 +1,10 @@
 import React, { MouseEventHandler } from "react";
 import styled from "styled-components";
 import Text from "../Text/Text";
-import { CommonButtonTagProps } from "../Common.types";
+import {
+  CommonButtonTagProps,
+  StyledCommonButtonTagProps,
+} from "../Common.types";
 
 export interface ButtonProps extends CommonButtonTagProps {
   hoverColor?: string;
@@ -9,30 +12,35 @@ export interface ButtonProps extends CommonButtonTagProps {
   onClick?: MouseEventHandler<HTMLButtonElement>;
 }
 
-export const StyledButton = styled.button<ButtonProps>`
+export interface StyledButtonProps extends StyledCommonButtonTagProps {
+  $hoverColor?: string;
+  $disabled?: boolean;
+}
+
+export const StyledButton = styled.button<StyledButtonProps>`
   border: none;
   cursor: pointer;
   padding: ${(props) =>
-    props.padding
-      ? props.padding
-      : props.size === "large"
+    props.$padding
+      ? props.$padding
+      : props.$size === "large"
         ? "15px 20px"
-        : props.size === "medium"
+        : props.$size === "medium"
           ? "10px 12px"
           : "7px"};
-  background-color: ${(props) => props.backgroundColor || `#d3d3d3`};
-  pointer-events: ${(props) => (props.disabled ? `none` : `auto`)};
-  opacity: ${(props) => (props.disabled ? 0.5 : 1)};
+  background-color: ${(props) => props.$backgroundColor || `#d3d3d3`};
+  pointer-events: ${(props) => (props.$disabled ? `none` : `auto`)};
+  opacity: ${(props) => (props.$disabled ? 0.5 : 1)};
   border-radius: ${(props) =>
-    props.borderRadius
-      ? props.borderRadius
-      : props.format === "rounded"
+    props.$borderRadius
+      ? props.$borderRadius
+      : props.$format === "rounded"
         ? "30px"
-        : props.format === "semiRounded"
+        : props.$format === "semiRounded"
           ? "5px"
           : "0px"};
   &:hover {
-    background-color: ${(props) => props.hoverColor || "#a9a9a9"};
+    background-color: ${(props) => props.$hoverColor || "#a9a9a9"};
   }
 `;
 
@@ -53,14 +61,13 @@ const Button = ({
   onClick,
 }: ButtonProps) => (
   <StyledButton
-    text={text}
-    backgroundColor={backgroundColor}
-    hoverColor={hoverColor}
-    format={format}
-    borderRadius={borderRadius}
-    size={size}
-    padding={padding}
-    disabled={disabled}
+    $backgroundColor={backgroundColor}
+    $hoverColor={hoverColor}
+    $format={format}
+    $borderRadius={borderRadius}
+    $size={size}
+    $padding={padding}
+    $disabled={disabled}
     onClick={onClick}
   >
     <Text