+++ src/components/base/Common.types.ts
@@ -7,6 +7,14 @@ export interface CommonTextProps {
   fontFamily?: string;
 }
 
+export interface StyledCommonTextProps {
+  $color?: string;
+  $weight?: "normal" | "bold";
+  $fontWeight?: number;
+  $fontSize?: string;
+  $fontFamily?: string;
+}
+
 export interface CommonButtonTagProps {
   text: string;
   textColor?: string;
@@ -21,6 +29,23 @@ export interface CommonButtonTagProps {
   padding?: string;
 }
 
+export interface StyledCommonButtonTagProps {
+  $textColor?: string;
+  $textWeight?: "normal" | "bold";
+  $textFontWeight?: number;
+  $textFontSize?: string;
+  $textFontFamily?: string;
+  $backgroundColor?: string;
+  $format?: "default" | "semiRounded" | "rounded";
+  $borderRadius?: string;
+  $size?: "small" | "medium" | "large";
+  $padding?: string;
+}
+
 export interface CommonTagProps extends CommonButtonTagProps {
   type?: "default" | "success" | "alert" | "error";
+}
+
+export interface StyledCommonTagProps extends StyledCommonButtonTagProps {
+  $type?: "default" | "success" | "alert" | "error";
 }