+++ src/components/base/Div/Div.tsx
@@ -13,19 +13,30 @@ export interface DivProps {
   gap?: string;
 }
 
-export const StyledDiv = styled.div<DivProps>`
-  ${(props) => props.margin && `margin: ${props.margin}`};
-  ${(props) => props.padding && `padding: ${props.padding}`};
-  ${(props) => props.height && `height: ${props.height}`};
-  ${(props) => props.width && `width: ${props.width}`};
+export interface StyledDivProps {
+  $margin?: string;
+  $padding?: string;
+  $height?: string;
+  $width?: string;
+  $alignContent?: string;
+  $alignItems?: string;
+  $justifyContent?: string;
+  $gap?: string;
+}
+
+export const StyledDiv = styled.div<StyledDivProps>`
+  ${(props) => props.$margin && `margin: ${props.$margin}`};
+  ${(props) => props.$padding && `padding: ${props.$padding}`};
+  ${(props) => props.$height && `height: ${props.$height}`};
+  ${(props) => props.$width && `width: ${props.$width}`};
   ${(props) =>
-    (props.alignContent || props.justifyContent || props.alignItems) &&
+    (props.$alignContent || props.$justifyContent || props.$alignItems) &&
     `display: flex`};
-  ${(props) => props.alignContent && `align-content: ${props.alignContent}`};
-  ${(props) => props.alignItems && `align-items: ${props.alignItems}`};
+  ${(props) => props.$alignContent && `align-content: ${props.$alignContent}`};
+  ${(props) => props.$alignItems && `align-items: ${props.$alignItems}`};
   ${(props) =>
-    props.justifyContent && `justify-content: ${props.justifyContent}`};
-  ${(props) => props.gap && `gap: ${props.gap}`};
+    props.$justifyContent && `justify-content: ${props.$justifyContent}`};
+  ${(props) => props.$gap && `gap: ${props.$gap}`};
 `;
 
 const Div = ({
@@ -40,14 +51,14 @@ const Div = ({
   gap,
 }: DivProps) => (
   <StyledDiv
-    margin={margin}
-    padding={padding}
-    height={height}
-    width={width}
-    alignContent={alignContent}
-    alignItems={alignItems}
-    justifyContent={justifyContent}
-    gap={gap}
+    $margin={margin}
+    $padding={padding}
+    $height={height}
+    $width={width}
+    $alignContent={alignContent}
+    $alignItems={alignItems}
+    $justifyContent={justifyContent}
+    $gap={gap}
   >
     {children}
   </StyledDiv>