+++ src/components/base/Img/Img.tsx
@@ -10,15 +10,22 @@ export interface ImgProps {
   borderRadius?: string;
 }
 
-export const StyledImg = styled.img<ImgProps>`
-  height: ${(props) => props.height || "auto"};
-  width: ${(props) => props.width || "auto"};
+export interface StyledImgProps {
+  $height?: string;
+  $width?: string;
+  $format?: string;
+  $borderRadius?: string;
+}
+
+export const StyledImg = styled.img<StyledImgProps>`
+  height: ${(props) => props.$height || "auto"};
+  width: ${(props) => props.$width || "auto"};
   border-radius: ${(props) =>
-    props.borderRadius
-      ? props.borderRadius
-      : props.format === "rounded"
+    props.$borderRadius
+      ? props.$borderRadius
+      : props.$format === "rounded"
         ? "30px"
-        : props.format === "semiRounded"
+        : props.$format === "semiRounded"
           ? "5px"
           : "0px"};
 `;
@@ -27,10 +34,10 @@ const Img = ({ src, alt, height, width, format, borderRadius }: ImgProps) => (
   <StyledImg
     src={src}
     alt={alt}
-    height={height}
-    width={width}
-    format={format}
-    borderRadius={borderRadius}
+    $height={height}
+    $width={width}
+    $format={format}
+    $borderRadius={borderRadius}
   />
 );