+++ src/components/base/Tag/Tag.tsx
@@ -1,35 +1,35 @@
 import React from "react";
 import styled from "styled-components";
 import Text from "../Text/Text";
-import { CommonTagProps } from "../Common.types";
+import { CommonTagProps, StyledCommonTagProps } from "../Common.types";
 
-export const StyledTag = styled.div<CommonTagProps>`
+export const StyledTag = styled.div<StyledCommonTagProps>`
   border: none;
   padding: ${(props) =>
-    props.padding
-      ? props.padding
-      : props.size === "large"
+    props.$padding
+      ? props.$padding
+      : props.$size === "large"
         ? "15px 20px"
-        : props.size === "medium"
+        : props.$size === "medium"
           ? "10px 12px"
           : "7px"};
   background-color: ${(props) =>
-    props.backgroundColor
-      ? props.backgroundColor
-      : props.type === "error"
+    props.$backgroundColor
+      ? props.$backgroundColor
+      : props.$type === "error"
         ? "#e97451"
-        : props.type === "alert"
+        : props.$type === "alert"
           ? "#f8de7e"
-          : props.type === "success"
+          : props.$type === "success"
             ? "#50c878"
             : "#d3d3d3"};
   pointer-events: none;
   border-radius: ${(props) =>
-    props.borderRadius
-      ? props.borderRadius
-      : props.format === "rounded"
+    props.$borderRadius
+      ? props.$borderRadius
+      : props.$format === "rounded"
         ? "30px"
-        : props.format === "semiRounded"
+        : props.$format === "semiRounded"
           ? "5px"
           : "0"};
   width: fit-content;
@@ -51,13 +51,12 @@ const Tag = ({
 }: CommonTagProps) => (
   <StyledTag
     data-testid="tag"
-    text={text}
-    type={type}
-    backgroundColor={backgroundColor}
-    format={format}
-    borderRadius={borderRadius}
-    size={size}
-    padding={padding}
+    $type={type}
+    $backgroundColor={backgroundColor}
+    $format={format}
+    $borderRadius={borderRadius}
+    $size={size}
+    $padding={padding}
   >
     <Text
       color={textColor || "#fff"}