+++ src/components/base/Text/Text.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import styled from "styled-components";
-import { CommonTextProps } from "../Common.types";
+import { CommonTextProps, StyledCommonTextProps } from "../Common.types";
 
-export const StyledText = styled.span<CommonTextProps>`
-  color: ${(props) => (props.color ? props.color : "#000")};
-  font-size: ${(props) => (props.fontSize ? props.fontSize : "16px")};
+export const StyledText = styled.span<StyledCommonTextProps>`
+  color: ${(props) => (props.$color ? props.$color : "#000")};
+  font-size: ${(props) => (props.$fontSize ? props.$fontSize : "16px")};
   font-weight: ${(props) =>
-    props.fontWeight
-      ? props.fontWeight
-      : props.weight
-        ? props.weight
+    props.$fontWeight
+      ? props.$fontWeight
+      : props.$weight
+        ? props.$weight
         : "normal"};
-  font-family: ${(props) => (props.fontFamily ? props.fontFamily : "Arial")};
+  font-family: ${(props) => (props.$fontFamily ? props.$fontFamily : "Arial")};
 `;
 
 const Text = ({
@@ -23,11 +23,11 @@ const Text = ({
   fontFamily,
 }: CommonTextProps) => (
   <StyledText
-    color={color}
-    weight={weight}
-    fontWeight={fontWeight}
-    fontSize={fontSize}
-    fontFamily={fontFamily}
+    $color={color}
+    $weight={weight}
+    $fontWeight={fontWeight}
+    $fontSize={fontSize}
+    $fontFamily={fontFamily}
   >
     {children}
   </StyledText>